Move format string checks to core.h
